DRA: allow attaching DRA objects to NodeInfo and PodInfo in scheduler framework
This allows components using the framework to do scheduling simulations
to simulate the DRA objects along Pods and Nodes.

Scheduler and possibly other places need to be updated to set the
DRA objects in NodeInfos wherever Nodes are set and the NodeInfo could
reach the DRA plugin. Left TODOs in possible places.

Then the DRA scheduler plugin needs to be updated to take the DRA objects
from NodeInfo/PodInfo instead of the live listers.

We still also need a way for the DRA scheduler plugin to reflect
changes to some of the objects caused by allocation back to the
NodeInfos/PodInfos.